// api/coreService/types/dto/PrayerTimesResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type DtoPrayerTimesResponse = {
    /**
     * @type string | undefined
    */
    asr?: string;
    /**
     * @type string | undefined
    */
    date?: string;
    /**
     * @type string | undefined
    */
    dhuhr?: string;
    /**
     * @type string | undefined
    */
    fajr?: string;
    /**
     * @type string | undefined
    */
    isha?: string;
    /**
     * @type string | undefined
    */
    maghrib?: string;
    /**
     * @type string | undefined
    */
    source?: string;
    /**
     * @type string | undefined
    */
    sunrise?: string;
    /**
     * @type string | undefined
    */
    timezone?: string;
};